[refactor/#5] data:onboard, data:onboard-api 모듈 분리
