--- core/data/build.gradle.kts
@@ -1,12 +1,0 @@
-plugins {
-    id("qriz.android.library")
-}
-
-android {
-    namespace = "com.qriz.app.core.data"
-}
-
-dependencies {
-    implementation(projects.core.network)
-    implementation(projects.core.datastore)
-}
