+++ core/data/onboard/onboard-api/build.gradle.kts
@@ -1,0 +1,10 @@
+plugins {
+    id("qriz.android.library")
+}
+
+android {
+    namespace = "com.qriz.app.core.data.onboard.onboard_api"
+}
+
+dependencies {
+}
