+++ core/data/onboard/onboard-api/src/main/java/com/qriz/app/core/data/onboard/onboard_api/model/Question.kt
@@ -1,0 +1,8 @@
+package com.qriz.app.core.data.onboard.onboard_api.model
+
+data class Question(
+    val id: Long,
+    val question: String,
+    val options: List<String>,
+    val timeLimit: Int,
+)
