--- core/data/src/main/java/com/qriz/app/core/data/model/test/Question.kt
@@ -1,8 +1,0 @@
-package com.qriz.app.core.data.model.test
-
-data class Question(
-    val id: Long,
-    val question: String,
-    val options: List<String>,
-    val timeLimit: Int,
-)
