+++ feature/onboard/build.gradle.kts
@@ -7,4 +7,5 @@ android {
 }
 
 dependencies {
+    implementation(projects.core.data.onboard.onboardApi)
 }
