+++ feature/onboard/src/main/java/com/qriz/app/feature/onboard/ui/screen/preview/PreviewScreen.kt
@@ -26,7 +26,7 @@ import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.qriz.app.core.data.model.test.Question
+import com.qriz.app.core.data.onboard.onboard_api.model.Question
 import com.qriz.app.core.designsystem.component.NavigationType
 import com.qriz.app.core.designsystem.component.QrizButton
 import com.qriz.app.core.designsystem.component.QrizTopBar
