+++ settings.gradle.kts
@@ -33,6 +33,8 @@ include(":core:network:user")
 
 include(":core:data:user:user-api")
 include(":core:data:user:user")
+include(":core:data:onboard:onboard-api")
+include(":core:data:onboard:onboard")
 
 include(":feature:main")
 include(":feature:splash")
